requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.3
blinker==1.9.0

# Production WSGI server
gunicorn==21.2.0

# Email validation dependencies
dnspython==2.4.2
requests==2.31.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0

# Environment variables
python-dotenv==1.0.0
